refactor: add new Subject List model in HomeViewModel because the hierarchy
