+++ app/src/main/java/com/example/savestudents/repository/HomeRepository.kt
@@ -2,7 +2,9 @@ package com.example.savestudents.repository
 
 import com.example.savestudents.constants.FirestoreDbConstants
 import com.example.savestudents.dto.SubjectListDto
-import com.example.savestudents.model.IHomeRepository
+import com.example.savestudents.dto.asDomainModel
+import com.example.savestudents.model.SubjectList
+import com.example.savestudents.model.repository.IHomeRepository
 import com.example.savestudents.service.FirebaseClient
 import com.example.savestudents.service.model.FirebaseResponseModel
 import com.example.savestudents.service.model.OnFailureModel
