+++ app/src/main/java/com/example/savestudents/viewModel/HomeViewModel.kt
@@ -4,7 +4,9 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.savestudents.dto.SubjectListDto
-import com.example.savestudents.model.IHomeRepository
+import com.example.savestudents.dto.asDomainModel
+import com.example.savestudents.model.SubjectList
+import com.example.savestudents.model.repository.IHomeRepository
 import com.example.savestudents.repository.HomeRepository
 import com.example.savestudents.service.model.FirebaseResponseModel
 import com.example.savestudents.service.model.OnFailureModel
@@ -12,13 +14,13 @@ import com.example.savestudents.service.model.OnFailureModel
 class HomeViewModel() : ViewModel() {
     private val repository: IHomeRepository = HomeRepository()
 
-    private var mSubjectList = MutableLiveData<List<SubjectListDto>>()
-    val subjectList: LiveData<List<SubjectListDto>> = mSubjectList
+    private var mSubjectList = MutableLiveData<List<SubjectList>>()
+    val subjectList: LiveData<List<SubjectList>> = mSubjectList
 
     fun getSubjectList() {
         repository.getSubjectList(object : FirebaseResponseModel<List<SubjectListDto>> {
             override fun onSuccess(model: List<SubjectListDto>) {
-                mSubjectList.value = model
+                mSubjectList.value = model.asDomainModel()
             }
 
             override fun onFailure(error: OnFailureModel) {
